Expose FXOutput
